feat: create AutomobilesController
